+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Change Log
 This file tracks released versions with the changes made to this project.
 
-## [Version 1.0.11](https://github.com/SocketMobile/swift-package-capturesdk/releases/tag/1.4.57)
+## [Version 1.0.9](https://github.com/SocketMobile/swift-package-maraca/releases/tag/1.0.9)
 ### New
-  * Maraca is now a [Swift Package Manager](https://github.com/SocketMobile/swift-package-capturesdk) 🎉🎉🎉
-  * Now contains CaptureSDK 1.5.2 as Swift package Manager too 🎉🎉🎉
+  * Maraca is now a [Swift Package Manager](https://github.com/SocketMobile/swift-package-maraca) 🎉🎉🎉
+  * Now contains [CaptureSDK 1.5.2](https://github.com/SocketMobile/swift-package-capturesdk/releases/tag/1.5.2) as a dependency
 
 ### Improvements
   * n/a